ACTIN-1862: Clean up some trial generator code
